TodolistResponse

+++ src/main/java/booungyi/todolist/todolist/TodolistResponse.java
@@ -6,4 +6,4 @@ import java.util.List;
 
 public record TodolistResponse(
         String todolistname,
-        Long count) { }
+        int count) { }
